refactor(log): move log to internal/logging package
